# ext/hivemq-edge-openapi-2023.9.yaml
openapi: 3.0.1
info:
  contact:
    url: https://www.hivemq.com
  description: |
    # Introduction
    HiveMQ Edge's REST API provides endpoints for the following use cases:
    - Authentication
    - Health Checking
    - Exploring Configuration
    ## Errors
    Conventional HTTP response codes are used to indicate the success or failure of an API request. Codes in the 2xx range generally indicate success. Codes in the 4xx range indicate an error that failed given the information provided (e.g., a required parameter was omitted). Codes in the 5xx range indicate an error on the server side.
    For all errors a JSON response with additional details is returned in the format [Problem JSON](https://tools.ietf.org/html/rfc7807).
    ## OpenAPI
    HiveMQ's REST API provides an OpenAPI 3.0 schema definition that can imported into popular API tooling (e.g. Postman) or can be used to generate client-code for multiple programming languages.
  title: HiveMQ Edge REST API
  version: "2023.9"
  x-logo:
    url: https://www.hivemq.com/img/svg/hivemq-bee.svg
...
paths:
...
  /api/v1/management/events:
    get:
      description: Get all bridges configured in the system.
      operationId: getEvents
      parameters:
        - description: Obtain all events since the specified epoch.
          in: query
          name: limit
          schema:
            type: integer
            format: int32
            default: 100
        - description: Obtain all events since the specified epoch.
          in: query
          name: since
          schema:
            type: integer
            format: int64
      responses:
        "200":
          content:
            application/json:
              examples:
                event-list-result:
                  description: Example response with several events.
                  summary: Event List result
                  value: {}
              schema:
                $ref: '#/components/schemas/EventList'
          description: Success
      summary: List most recent events in the system
      tags:
        - Events
...
components:
  schemas:
...
    Event:
      type: object
      description: List of result items that are returned by this endpoint
      properties:
        associatedObject:
          $ref: '#/components/schemas/TypeIdentifier'
        created:
          type: string
          format: date-time
          description: Time the event was in date format
        identifier:
          $ref: '#/components/schemas/TypeIdentifier'
        message:
          type: string
          description: The message associated with the event. A message will be no
            more than 1024 characters in length
        payload:
          $ref: '#/components/schemas/Payload'
        severity:
          type: string
          description: The severity that this log is considered to be
          enum:
            - INFO
            - WARN
            - ERROR
            - CRITICAL
        source:
          $ref: '#/components/schemas/TypeIdentifier'
        timestamp:
          type: integer
          format: int64
          description: Time the event was generated in epoch format
      required:
        - created
        - identifier
        - message
        - severity
        - timestamp
    EventList:
      type: object
      properties:
        items:
          type: array
          description: List of result items that are returned by this endpoint
          items:
            $ref: '#/components/schemas/Event'
...
    Payload:
      type: object
      description: Object to denote the payload of the event
      properties:
        content:
          type: string
          description: The content of the payload encoded as a string
        contentType:
          type: string
          description: The content type of the payload that the event contains
          enum:
            - JSON
            - PLAIN_TEXT
            - XML
            - CSV
      required:
        - contentType
...
    TypeIdentifier:
      type: object
      description: The type-identifier of the object who caused the event to be generated
      properties:
        fullQualifiedIdentifier:
          type: string
        identifier:
          type: string
          description: "The identifier associated with the object, a combination of\
            \ type and identifier is used to uniquely identify an object in the system"
        type:
          type: string
          description: The type of the associated object/entity
          enum:
            - BRIDGE
            - ADAPTER
            - ADAPTER_TYPE
            - EVENT
            - USER
      required:
        - type
...